// Frontend/src/app/model/kampagne.ts
export class Kampagne {
    id: number;
    name: string;
    TBL: string;
    CAMPAIGN_ID: string;
    SERVICE: string;
    TICKET: string;
    TS_CREATION: string;
    REPRO_SPERRE: string;
    REPRO_EM: string;
    REPRO_RUFNR: string;
    REPRO_PARAM: string;
    STATUS: string;
    ANZ_JOBS: string;
    DELAY: string;
    WAITING_PERIOD: string;

}